Added a new Signal type and support for fixed point arithmetic.
The Test file has been modified to use the new Signal type and to show
an example of use of the new Fixed type.

The fixed point type is following the ITU G.191 spec commonly used
for audio and speech development.

The plotting function have been modified to use existential types which allow
the mixing of signal of different types.