Generic implementation for fixed point
The fixed point implementation is now totally generic on the fractional part length and the
saturation mode. Only the data type for the base value needs class instantiation. It is done through
CPP macros.

The class SuperInt has been renamed to NumberInfo.

Rounding has been removed since it will be supported as a new type field on the fixed point number.