Inline, audio, new plot
Added inline pragmas for the fixed point
Added an export to wav format for testing
The plot has been modified a lot.
Now, a displayable is a more complex data type because of the need to generate the resources (like images)
in the PDF monad. So, a displayable must have some part in PDF and some parts in Draw.

When a drawing is too complex, the PDF viewer will collapse. So, to avoid this, the plotting function are switching
to a raster rendering for the curves (only the curves : other parts are still PDF and vectorial).
The rendered image is included in the PDF. Rendering is using a simple Bresenham algorithm with no aliasing.

It is still a bit slow for complex pictures (example with 64 000 samples).

But at least something is displayed. It was not possible with a PDF document containing 64000 lines.